Fix Start auth server entry split
