+++ .changeset/convex-auth-provider-query-client.md
@@ -9,4 +9,4 @@
 ## Patches
 
 - Keep the existing auth store attached when reusing a Convex query client before `ConvexAuthProvider` resyncs it.
-- Fix `kitcn/auth/start` so Nitro production builds can trace and include the TanStack Start server dependency.
+- Fix `kitcn/auth/start/server` so Nitro production builds can trace and include the TanStack Start server dependency without making `kitcn/auth/start` unsafe for browser loaders.
