+++ packages/kitcn/package.json
@@ -28,6 +28,7 @@
     "./auth/http": "./dist/auth/http/index.js",
     "./auth/nextjs": "./dist/auth/nextjs/index.js",
     "./auth/start": "./dist/auth/start/index.js",
+    "./auth/start/server": "./dist/auth/start/server/index.js",
     "./crpc": "./dist/crpc/index.js",
     "./orm": "./dist/orm/index.js",
     "./plugins": "./dist/plugins/index.js",
