+++ www/content/docs/tanstack-start.mdx
@@ -107,7 +107,7 @@ export const {
 ### Auth Server
 
 ```ts showLineNumbers title="src/lib/convex/auth-server.ts"
-import { convexBetterAuthReactStart } from 'kitcn/auth/start';
+import { convexBetterAuthReactStart } from 'kitcn/auth/start/server';
 
 export const {
   handler,
